Import KeysView, ValuesView from collections.abc to unblock upgrade past Python 3.6
When depending on this library, it's currently not possible to upgrade past Python 3.6 as KeysView, ValuesView and many other things are no longer exposed via `collections` but via `collections.abc` (see: https://github.com/python/cpython/issues/70176). This change is backwards compatible and is tested to work from Python 3.6.* to Python 3.10.* at least.